+++ src/main/java/jackiecrazy/wardance/skill/grapple/Wrestle.java
@@ -54,7 +54,7 @@ public class Wrestle extends Skill {
     }
 
     @Override
-    public float spiritConsumption(LivingEntity caster) {
+    public int spiritConsumption(LivingEntity caster) {
         return 1;
     }
